cosmetic: rename addressIdentity.js to address-identity.js
